fastapi_app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
odmantic==0.9.2
pydantic==1.10.7
python-dotenv==1.0.0
email-validator==2.0.0.post2
